Extracted functions out of main()
Main body of script now is split into several concise and localized
functions. We now also use custom exceptions to group all error handling
and returns into its own area. So the main is easier to read as there is
no extra checks and the like. Not completely sure if all of the
exceptins are for "Exceptional situations," but it makes the code a lot
cleaner.

Note this is one step on refactoring. Still need to complete the
exception definitions and their use (passing of info.)