added number of times to try updating option
This way an user can specify how many tries, so if apt-get update is
failing regularly enough they can just specify multiple tries to
increase the probability of a successful update. It also allows for the
user to specify 0 updates, in which case it won't attempt updating.

Added no-op ordinal() function to allow for better log printing of which
try just failed. Can go back later and implement should we so desire.