added network_check option
Network check is now disabled by default and can be enabled via the new
command line option. We're doing this b/c network check has frequently
been incorrect and doesn't add much of anything to the program in terms
of benefits. We're leaving it in for now until either I get sick of it
or we get some feedback (former more likely to happen.) I want to note
that the number of tries to update option which is in testing as I write
this is more useful, and in my opinion, makes the update check
unnecessary.