added sleep_period
Can now sleep/wait in between update tries. This can be useful in if the
update fail is due to certain circumstances. E.g: if the system runs
apt-get/dpkg/etc every so often for a period of time. By making it wait
before trying to update again, we might end up trying after the other
program has exited and released the lock on dpkg. Also, it prevents this
program from spamming update.