added option: time_format
Allows for user to specify a custom time format to be used for the
placeholder in the template. According to man strftime, %c is the
prefered default time format, so that's the default I choose here. Plus
its what has been used by asctime() (no pararms), so it must be a
reasonable default.